Rename #[trait_alias] to #[blanket_trait]
