+++ crates/cgp-component-macro-lib/src/blanket_trait/mod.rs
