+++ crates/cgp-component-macro-lib/src/blanket_trait/remove_self_path.rs
