+++ crates/cgp-component-macro-lib/src/entrypoints/mod.rs
@@ -1,3 +1,4 @@
+mod blanket_trait;
 mod cgp_auto_getter;
 mod cgp_component;
 mod cgp_context;
@@ -10,8 +11,8 @@ mod define_preset;
 mod delegate_components;
 mod re_export_imports;
 mod replace_with;
-mod trait_alias;
 
+pub use blanket_trait::*;
 pub use cgp_auto_getter::*;
 pub use cgp_component::*;
 pub use cgp_context::*;
@@ -24,4 +25,3 @@ pub use define_preset::*;
 pub use delegate_components::*;
 pub use re_export_imports::*;
 pub use replace_with::*;
-pub use trait_alias::*;
