+++ crates/cgp-component-macro-lib/src/tests/mod.rs
@@ -1,3 +1,4 @@
+pub mod blanket_trait;
 pub mod check_components;
 pub mod define_preset;
 pub mod delegate_components;
@@ -7,4 +8,3 @@ pub mod derive_getter;
 pub mod derive_provider;
 pub mod for_each_replace;
 pub mod helper;
-pub mod trait_alias;
