+++ crates/cgp-error/src/traits/has_async_error_type.rs
@@ -1,14 +1,14 @@
 use cgp_async::Async;
-use cgp_component_macro::trait_alias;
+use cgp_component_macro::blanket_trait;
 
 use crate::traits::HasErrorType;
 use crate::{CanRaiseError, CanWrapError};
 
-#[trait_alias]
+#[blanket_trait]
 pub trait HasAsyncErrorType: Async + HasErrorType<Error: Async> {}
 
-#[trait_alias]
+#[blanket_trait]
 pub trait CanRaiseAsyncError<SourceError>: HasAsyncErrorType + CanRaiseError<SourceError> {}
 
-#[trait_alias]
+#[blanket_trait]
 pub trait CanWrapAsyncError<SourceError>: HasAsyncErrorType + CanWrapError<SourceError> {}
